+++ spring-platform-cloud-common/src/main/java/com/jtang/common/annotation/OperationLog.java
@@ -11,5 +11,14 @@ import java.lang.annotation.*;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface OperationLog {
 
+    /** 模快名:默认为类注解@Api的值 */
+    String module() default "";
+
+    /** 记录日志 是：true 否：false */
+    boolean record() default true;
+
+    /** 敏感字段信息 */
+    String[] sensitive() default {};
+
 
 }
