+++ spring-platform-cloud-common/src/main/java/com/jtang/common/utils/HttpUtils.java
@@ -1,180 +1,31 @@
 package com.jtang.common.utils;
 
-import org.springframework.http.*;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.context.request.RequestContextHolder;
-import org.springframework.web.context.request.ServletRequestAttributes;
-
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 /**
- * 网络请求工具类
- * @date 2020/7/4 12:21
- * @author LinJinTang
+ * HTTP 请求信息工具类
+ * @author linjt
+ * @date 2020/8/7
  */
 public class HttpUtils {
 
-    /**
-     * get请求
-     *
-     * @param url 请求地址
-     * @param params 请求参数
-     * @return String
-     */
-    public static String get(String url, MultiValueMap<String, String> params) {
-        return get(url, params, null);
-    }
-
-    /**
-     * get请求
-     *
-     * @param url 请求地址
-     * @param params  请求参数
-     * @param headers 请求头
-     * @return String
-     */
-    public static String get(String url, MultiValueMap<String, String> params, MultiValueMap<String, String> headers) {
-        return request(url, params, headers, HttpMethod.GET);
-    }
-
-    /**
-     * post请求
-     *
-     * @param url 请求地址
-     * @param params 请求参数
-     * @return String
-     */
-    public static String post(String url, MultiValueMap<String, String> params) {
-        return post(url, params, null);
-    }
-
-    /**
-     * post请求
-     *
-     * @param url 请求地址
-     * @param params  请求参数
-     * @param headers 请求头
-     * @return String
-     */
-    public static String post(String url, MultiValueMap<String, String> params, MultiValueMap<String, String> headers) {
-        return request(url, params, headers, HttpMethod.POST);
-    }
-
-    /**
-     * put请求
-     *
-     * @param url 请求地址
-     * @param params 请求参数
-     * @return String
-     */
-    public static String put(String url, MultiValueMap<String, String> params) {
-        return put(url, params, null);
-    }
-
-    /**
-     * put请求
-     *
-     * @param url 请求地址
-     * @param params  请求参数
-     * @param headers 请求头
-     * @return String
-     */
-    public static String put(String url, MultiValueMap<String, String> params, MultiValueMap<String, String> headers) {
-        return request(url, params, headers, HttpMethod.PUT);
-    }
-
-    /**
-     * delete请求
-     *
-     * @param url 请求地址
-     * @param params 请求参数
-     * @return String
-     */
-    public static String delete(String url, MultiValueMap<String, String> params) {
-        return delete(url, params, null);
-    }
-
-    /**
-     * delete请求
-     *
-     * @param url 请求地址
-     * @param params  请求参数
-     * @param headers 请求头
-     * @return String
-     */
-    public static String delete(String url, MultiValueMap<String, String> params, MultiValueMap<String, String> headers) {
-        return request(url, params, headers, HttpMethod.DELETE);
-    }
-
-    /**
-     * 表单请求
-     *
-     * @param url 请求地址
-     * @param params  请求参数
-     * @param headers 请求头
-     * @param method  请求方式
-     * @return String
-     */
-    public static String request(String url, MultiValueMap<String, String> params, MultiValueMap<String, String> headers, HttpMethod method) {
-        if (params == null) {
-            params = new LinkedMultiValueMap<>();
+    public static String getIpAddr(HttpServletRequest request) {
+        String ip = request.getHeader("x-forwarded-for");
+        if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
+            ip = request.getHeader("Proxy-Client-IP");
         }
-        return request(url, params, headers, method, MediaType.APPLICATION_FORM_URLENCODED);
-    }
-
-    /**
-     * http请求
-     *
-     * @param url 请求地址
-     * @param params    请求参数
-     * @param headers   请求头
-     * @param method    请求方式
-     * @param mediaType 参数类型
-     * @return String
-     */
-    public static String request(String url, Object params, MultiValueMap<String, String> headers, HttpMethod method, MediaType mediaType) {
-        if (url == null || url.trim().isEmpty()) {
-            return null;
+        if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
+            ip = request.getHeader("WL-Proxy-Client-IP");
         }
-        RestTemplate client = new RestTemplate();
-        // header
-        HttpHeaders httpHeaders = new HttpHeaders();
-        if (headers != null) {
-            httpHeaders.addAll(headers);
+        if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
+            ip = request.getHeader("HTTP_CLIENT_IP");
         }
-        // 提交方式：表单、json
-        httpHeaders.setContentType(mediaType);
-        HttpEntity<Object> httpEntity = new HttpEntity(params, httpHeaders);
-        ResponseEntity<String> response = client.exchange(url, method, httpEntity, String.class);
-        return response.getBody();
-    }
-
-
-    /** 获取请求响应体 */
-    public static HttpServletResponse getResponse(){
-        HttpServletResponse response = getServletRequestAttributes().getResponse();
-        if(response == null){
-            throw new RuntimeException();
+        if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
+            ip = request.getHeader("HTTP_X_FORWARDED_FOR");
         }
-        return response;
-    }
-
-    /** 获取请求响应体 */
-    public static HttpServletRequest getRequest(){
-        return getServletRequestAttributes().getRequest();
-    }
-
-    /** 获取网络请求体 */
-    private static ServletRequestAttributes getServletRequestAttributes(){
-        ServletRequestAttributes servletRequestAttributes = ((ServletRequestAttributes) RequestContextHolder.getRequestAttributes());
-        if(servletRequestAttributes == null){
-            throw new RuntimeException();
+        if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
+            ip = request.getRemoteAddr();
         }
-        return servletRequestAttributes;
+        return ip;
     }
-
-
 }
