+++ spring-platform-cloud-feign/src/main/java/com/jtang/feign/OutFeignRequest.java
@@ -1,6 +1,6 @@
 package com.jtang.feign;
 
-import com.jtang.common.utils.HttpUtils;
+import com.jtang.common.utils.HttpRequestUtils;
 import feign.RequestInterceptor;
 import feign.RequestTemplate;
 import org.springframework.stereotype.Component;
@@ -17,7 +17,7 @@ public class OutFeignRequest implements RequestInterceptor {
 
     @Override
     public void apply(RequestTemplate requestTemplate) {
-        HttpServletRequest request = HttpUtils.getRequest();
+        HttpServletRequest request = HttpRequestUtils.getRequest();
         //添加token
         requestTemplate.header("Authorization", request.getHeader("Authorization"));
     }
